feat: LÓTS OF CHANGES

+++ new-server/app/Http/Controllers/BranchController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 class BranchController extends Controller {
 
@@ -14,7 +15,7 @@ class BranchController extends Controller {
         $store_id = Auth::guard('stores')->user()->id;
         $data = $request->all();
         $rules = [
-            'name' => ['required', 'string', 'max:255'],
+            'name' => ['required', 'string', 'max:255', Rule::unique('branches')->where('store_id', $store_id)],
             'address' => ['required', 'string', 'max:1000'],
             'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg', 'max:5012'],
         ];
@@ -28,24 +29,19 @@ class BranchController extends Controller {
             ], 400);
         }
 
-        // check if branch name within the same store is used
-        $branch = Branch::where('name', $data['name'])->where('store_id', $store_id)->first();
-        if ($branch) {
-            return response()->json([
-                'message' => 'Branch name already exists.',
-            ], 400);
-        }
-
         $isExist = $request->hasFile('image');
 
+        // replace space in branch name with underscore
+        $image_name = $store_id . str_replace(' ', '_', $data['name']);
+
         $path = $isExist
-            ? $request->file('image')->storeAs('branches', $store_id . $request->input('name') . "." . $request->file('image')->getClientOriginalExtension() )
+            ? $request->file('image')->storeAs('branches', $image_name . "." . $request->file('image')->getClientOriginalExtension() )
             : 'branches/default.jpg';
 
         $branch = Branch::create([
             'name' => $data['name'],
             'address' => $data['address'],
-            'store_id' => Auth::guard('stores')->user()->id,
+            'store_id' => $store_id,
             'image' => $path,
         ]);
 
@@ -53,8 +49,6 @@ class BranchController extends Controller {
     }
 
     public function getBranchImage($filePath) {
-        // $filePath = 'branches' . '/' . $fileName;
-        error_log("Got file: " . $filePath);
         if (!Storage::exists($filePath)) {
             return response()->json([
                 'message' => 'File not found.',
@@ -63,10 +57,18 @@ class BranchController extends Controller {
         return response()->file(storage_path('app' . DIRECTORY_SEPARATOR . $filePath));
     }
 
-    public function getBranches() {
+    public function getBranches(Request $request) {
         $store_id = Auth::guard('stores')->user()->id;
 
-        $branches = Branch::where('store_id', $store_id)->get();
+        $search_text = $request->query('search') ?? '';
+
+        // search branch by name, address
+        $branches = Branch::where('store_id', $store_id)
+            ->where(function ($query) use ($search_text) {
+                $query->where('name', 'like', '%' . $search_text . '%')
+                    ->orWhere('address', 'like', '%' . $search_text . '%');
+            })
+            ->get();
 
         return response()->json($branches);
     }
@@ -88,8 +90,12 @@ class BranchController extends Controller {
 
     public function update(Request $request, $branch_id) {
         $store_id = Auth::guard('stores')->user()->id;
+
         $data = $request->all();
+        error_log($branch_id);
+        $data['branch_id'] = $branch_id;
         $rules = [
+            'branch_id' => ['required', 'integer', Rule::exists('branches', 'id')->where('store_id', $store_id)],
             'name' => ['nullable', 'string', 'max:255'],
             'address' => ['nullable', 'string', 'max:1000'],
             'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg', 'max:5012'],
@@ -103,18 +109,15 @@ class BranchController extends Controller {
                 'errors' => $validator->errors(),
             ], 400);
         }
-        // make sure store owns branch
+
         $branch = Branch::where('store_id', $store_id)->where('id', $branch_id)->first();
-        if (!$branch) {
-            return response()->json([
-                'message' => 'Unauthorized.',
-            ], 401);
-        }
 
         $isExist = $request->hasFile('image');
+        // replace space in branch name with underscore
+        $image_name = $store_id . str_replace(' ', '_', $data['name']);
         $path = $isExist
-            ? $request->file('image')->storeAs('branches', $store_id . $request->input('name') . "." . $request->file('image')->getClientOriginalExtension() )
-            : 'branches/default.jpg';
+            ? $request->file('image')->storeAs('branches', $image_name . "." . $request->file('image')->getClientOriginalExtension() )
+            : $branch->image;
 
         $branch->name = $data['name'] ?? $branch->name;
         $branch->address = $data['address'] ?? $branch->address;
@@ -126,9 +129,13 @@ class BranchController extends Controller {
 
     public function delete(Request $request, $branch_id) {
         $store_id = Auth::guard('stores')->user()->id;
-        $data = $request->all();
         // make sure store owns branch
         $branch = Branch::where('store_id', $store_id)->where('id', $branch_id)->first();
+        if (!$branch) {
+            return response()->json([
+                'message' => 'Branch not found.',
+            ], 404);
+        }
         $branch->delete();
 
         return response()->json($branch);
